Compact index format to v6 with ~40% size reduction
- Switch NodeId from 64-bit to 32-bit (sufficient for ~4B nodes)
- Switch header offsets from 64-bit to 32-bit (4GB index limit)
- Use length-prefixed strings (u16) in string table, eliminating
  separate length fields in RawFileEntry and RawCommandEntry
- Add overflow protection for index size calculations
- Clean break from v5 (no backward compatibility)

Struct size changes:
- RawHeader: 64 → 40 bytes (37.5% smaller)
- RawFileEntry: 96 → 64 bytes (33% smaller)
- RawCommandEntry: 64 → 24 bytes (62.5% smaller)
- RawEdge: 24 → 16 bytes (33% smaller)

+++ include/pup/core/types.hpp
@@ -9,7 +9,8 @@
 namespace pup {
 
 /// Unique identifier for nodes in the dependency graph
-using NodeId = std::uint64_t;
+/// 32-bit is sufficient for ~4 billion nodes (even AOSP has < 10M)
+using NodeId = std::uint32_t;
 
 /// Invalid node ID sentinel
 inline constexpr auto INVALID_NODE_ID = NodeId { 0 };
